torment/requirements.txt
discord.py>=2.4.0,<3.0.0
discord-ios>=1.0.0
asyncpg>=0.29.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
psutil>=5.9.0
aiohttp>=3.9.0
Pillow>=10.0.0
colorthief>=0.2.1
requests>=2.31.0
humanize>=4.9.0
humanfriendly>=10.0
shazamio>=0.6.0
jishaku>=2.5.0
dateparser>=1.2.0
PyNaCl>=1.5.0
